+++ evmrpc/send.go
@@ -3,8 +3,10 @@ package evmrpc
 import (
 	"context"
 	"errors"
+	"fmt"
 	"time"
 
+	"github.com/cosmos/cosmos-sdk/baseapp"
 	"github.com/cosmos/cosmos-sdk/client"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
@@ -12,7 +14,9 @@ import (
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/lib/ethapi"
+	"github.com/ethereum/go-ethereum/rpc"
 	"github.com/ethereum/go-ethereum/signer/core/apitypes"
+	"github.com/sei-protocol/sei-chain/precompiles/wasmd"
 	"github.com/sei-protocol/sei-chain/x/evm/keeper"
 	"github.com/sei-protocol/sei-chain/x/evm/types"
 	"github.com/sei-protocol/sei-chain/x/evm/types/ethtx"
@@ -34,7 +38,8 @@ type SendConfig struct {
 	slow bool
 }
 
-func NewSendAPI(tmClient rpcclient.Client, txConfig client.TxConfig, sendConfig *SendConfig, k *keeper.Keeper, ctxProvider func(int64) sdk.Context, homeDir string, simulateConfig *SimulateConfig, connectionType ConnectionType) *SendAPI {
+func NewSendAPI(tmClient rpcclient.Client, txConfig client.TxConfig, sendConfig *SendConfig, k *keeper.Keeper, ctxProvider func(int64) sdk.Context, homeDir string, simulateConfig *SimulateConfig, app *baseapp.BaseApp,
+	antehandler sdk.AnteHandler, connectionType ConnectionType) *SendAPI {
 	return &SendAPI{
 		tmClient:       tmClient,
 		txConfig:       txConfig,
@@ -42,7 +47,7 @@ func NewSendAPI(tmClient rpcclient.Client, txConfig client.TxConfig, sendConfig 
 		keeper:         k,
 		ctxProvider:    ctxProvider,
 		homeDir:        homeDir,
-		backend:        NewBackend(ctxProvider, k, txConfig.TxDecoder(), tmClient, simulateConfig),
+		backend:        NewBackend(ctxProvider, k, txConfig.TxDecoder(), tmClient, simulateConfig, app, antehandler),
 		connectionType: connectionType,
 	}
 }
@@ -63,10 +68,17 @@ func (s *SendAPI) SendRawTransaction(ctx context.Context, input hexutil.Bytes) (
 	if err != nil {
 		return
 	}
+	gasUsedEstimate, err := s.simulateTx(ctx, tx)
+	if err != nil {
+		tx, _ = msg.AsTransaction()
+		gasUsedEstimate = tx.Gas() // if issue simulating, fallback to gas limit
+	}
 	txBuilder := s.txConfig.NewTxBuilder()
 	if err = txBuilder.SetMsgs(msg); err != nil {
 		return
 	}
+	fmt.Printf("DEBUG: [SendRawTransaction] setting gas estimate of %+v on tx builder\n", gasUsedEstimate)
+	txBuilder.SetGasEstimate(gasUsedEstimate)
 	txbz, encodeErr := s.txConfig.TxEncoder()(txBuilder.GetTx())
 	if encodeErr != nil {
 		return hash, encodeErr
@@ -92,6 +104,66 @@ func (s *SendAPI) SendRawTransaction(ctx context.Context, input hexutil.Bytes) (
 		}
 	}
 	return
+}
+
+func (s *SendAPI) simulateTx(ctx context.Context, tx *ethtypes.Transaction) (estimate uint64, err error) {
+	var from common.Address
+	if tx.Type() == ethtypes.DynamicFeeTxType {
+		signer := ethtypes.NewLondonSigner(s.keeper.ChainID(s.ctxProvider(LatestCtxHeight)))
+		from, err = signer.Sender(tx)
+		if err != nil {
+			err = fmt.Errorf("failed to get sender for dynamic fee tx: %w", err)
+			return
+		}
+	} else if tx.Protected() {
+		signer := ethtypes.NewEIP155Signer(s.keeper.ChainID(s.ctxProvider(LatestCtxHeight)))
+		from, err = signer.Sender(tx)
+		if err != nil {
+			err = fmt.Errorf("failed to get sender for protected tx: %w", err)
+			return
+		}
+	} else {
+		signer := ethtypes.HomesteadSigner{}
+		from, err = signer.Sender(tx)
+		if err != nil {
+			err = fmt.Errorf("failed to get sender for homestead tx: %w", err)
+			return
+		}
+	}
+	input_ := (hexutil.Bytes)(tx.Data())
+	gas_ := hexutil.Uint64(tx.Gas())
+	nonce_ := hexutil.Uint64(tx.Nonce())
+	al := tx.AccessList()
+	bNrOrHash := rpc.BlockNumberOrHashWithNumber(rpc.LatestBlockNumber)
+	ctx = context.WithValue(ctx, CtxIsWasmdPrecompileCallKey, wasmd.IsWasmdCall(tx.To()))
+	gp := tx.GasPrice()
+	maxFeePerGas := tx.GasFeeCap()
+	maxPriorityFeePerGas := tx.GasTipCap()
+	if gp != nil {
+		maxFeePerGas = nil
+		maxPriorityFeePerGas = nil
+	} else {
+		gp = nil
+	}
+	txArgs := ethapi.TransactionArgs{
+		From:                 &from,
+		To:                   tx.To(),
+		Gas:                  &gas_,
+		GasPrice:             (*hexutil.Big)(gp),
+		MaxFeePerGas:         (*hexutil.Big)(maxFeePerGas),
+		MaxPriorityFeePerGas: (*hexutil.Big)(maxPriorityFeePerGas),
+		Value:                (*hexutil.Big)(tx.Value()),
+		Nonce:                &nonce_,
+		Input:                &input_,
+		AccessList:           &al,
+		ChainID:              (*hexutil.Big)(tx.ChainId()),
+	}
+	estimate_, err := ethapi.DoEstimateGas(ctx, s.backend, txArgs, bNrOrHash, nil, s.backend.RPCGasCap())
+	if err != nil {
+		err = fmt.Errorf("failed to estimate gas: %w", err)
+		return
+	}
+	return uint64(estimate_), nil
 }
 
 func (s *SendAPI) SignTransaction(_ context.Context, args apitypes.SendTxArgs, _ *string) (result *ethapi.SignTransactionResult, returnErr error) {