+++ precompiles/json/json.go
@@ -17,9 +17,10 @@ import (
 )
 
 const (
-	ExtractAsBytesMethod     = "extractAsBytes"
-	ExtractAsBytesListMethod = "extractAsBytesList"
-	ExtractAsUint256Method   = "extractAsUint256"
+	ExtractAsBytesMethod          = "extractAsBytes"
+	ExtractAsBytesListMethod      = "extractAsBytesList"
+	ExtractAsUint256Method        = "extractAsUint256"
+	ExtractAsBytesFromArrayMethod = "extractAsBytesFromArray"
 )
 
 const JSONAddress = "0x0000000000000000000000000000000000001003"
@@ -31,9 +32,10 @@ const GasCostPerByte = 100 // TODO: parameterize
 var f embed.FS
 
 type PrecompileExecutor struct {
-	ExtractAsBytesID     []byte
-	ExtractAsBytesListID []byte
-	ExtractAsUint256ID   []byte
+	ExtractAsBytesID          []byte
+	ExtractAsBytesListID      []byte
+	ExtractAsUint256ID        []byte
+	ExtractAsBytesFromArrayID []byte
 }
 
 func NewPrecompile() (*pcommon.Precompile, error) {
@@ -49,6 +51,8 @@ func NewPrecompile() (*pcommon.Precompile, error) {
 			p.ExtractAsBytesListID = m.ID
 		case ExtractAsUint256Method:
 			p.ExtractAsUint256ID = m.ID
+		case ExtractAsBytesFromArrayMethod:
+			p.ExtractAsBytesFromArrayID = m.ID
 		}
 	}
 
@@ -79,6 +83,8 @@ func (p PrecompileExecutor) Execute(ctx sdk.Context, method *abi.Method, caller 
 
 		uint_.FillBytes(byteArr)
 		return byteArr, nil
+	case ExtractAsBytesFromArrayMethod:
+		return p.extractAsBytesFromArray(ctx, method, args, value)
 	}
 	return
 }
@@ -171,4 +177,39 @@ func (p PrecompileExecutor) ExtractAsUint256(_ sdk.Context, _ *abi.Method, args 
 	}
 
 	return value, nil
+}
+
+func (p PrecompileExecutor) extractAsBytesFromArray(_ sdk.Context, method *abi.Method, args []interface{}, value *big.Int) ([]byte, error) {
+	if err := pcommon.ValidateNonPayable(value); err != nil {
+		return nil, err
+	}
+
+	if err := pcommon.ValidateArgsLength(args, 2); err != nil {
+		return nil, err
+	}
+
+	// type assertion will always succeed because it's already validated in p.Prepare call in Run()
+	bz := args[0].([]byte)
+	var decoded []gjson.RawMessage
+	if err := gjson.Unmarshal(bz, &decoded); err != nil {
+		return nil, err
+	}
+	if len(decoded) > 1<<16 {
+		return nil, errors.New("input array is larger than 2^16")
+	}
+	index, ok := args[1].(uint16)
+	if !ok {
+		return nil, errors.New("index must be uint16")
+	}
+	if int(index) >= len(decoded) {
+		return nil, fmt.Errorf("index %d is out of bounds", index)
+	}
+	result := decoded[index]
+
+	// in the case of a string value, remove the quotes
+	if len(result) >= 2 && result[0] == '"' && result[len(result)-1] == '"' {
+		result = result[1 : len(result)-1]
+	}
+
+	return method.Outputs.Pack([]byte(result))
 }