+++ x/evm/keeper/msg_server.go
@@ -44,13 +44,7 @@ func NewMsgServerImpl(keeper *Keeper) types.MsgServer {
 
 var _ types.MsgServer = msgServer{}
 
-func (server msgServer) EVMTransaction(goCtx context.Context, msg *types.MsgEVMTransaction) (serverRes *types.MsgEVMTransactionResponse, err error) {
-	if msg.IsAssociateTx() {
-		// no-op in msg server for associate tx; all the work have been done in ante handler
-		return &types.MsgEVMTransactionResponse{}, nil
-	}
-	ctx := sdk.UnwrapSDKContext(goCtx)
-	tx, _ := msg.AsTransaction()
+func (k *Keeper) PrepareCtxForEVMTransaction(ctx sdk.Context, tx *ethtypes.Transaction) (sdk.Context, sdk.GasMeter) {
 	isWasmdPrecompileCall := wasmd.IsWasmdCall(tx.To())
 	if isWasmdPrecompileCall {
 		ctx = ctx.WithEVMEntryViaWasmdPrecompile(true)
@@ -64,6 +58,17 @@ func (server msgServer) EVMTransaction(goCtx context.Context, msg *types.MsgEVMT
 	// 	4. At the end of message server, gas consumed by EVM is adjusted to Sei's unit and counted in the original gas meter, because that original gas meter will be used to count towards block gas after message server returns
 	originalGasMeter := ctx.GasMeter()
 	ctx = ctx.WithGasMeter(sdk.NewInfiniteGasMeterWithMultiplier(ctx))
+	return ctx, originalGasMeter
+}
+
+func (server msgServer) EVMTransaction(goCtx context.Context, msg *types.MsgEVMTransaction) (serverRes *types.MsgEVMTransactionResponse, err error) {
+	if msg.IsAssociateTx() {
+		// no-op in msg server for associate tx; all the work have been done in ante handler
+		return &types.MsgEVMTransactionResponse{}, nil
+	}
+	ctx := sdk.UnwrapSDKContext(goCtx)
+	tx, _ := msg.AsTransaction()
+	ctx, originalGasMeter := server.PrepareCtxForEVMTransaction(ctx, tx)
 
 	stateDB := state.NewDBImpl(ctx, &server, false)
 	emsg := server.GetEVMMessage(ctx, tx, msg.Derived.SenderEVMAddr)
@@ -106,7 +111,7 @@ func (server msgServer) EVMTransaction(goCtx context.Context, msg *types.MsgEVMT
 			)
 			return
 		}
-		if isWasmdPrecompileCall {
+		if ctx.EVMEntryViaWasmdPrecompile() {
 			syntheticReceipt, err := server.GetTransientReceipt(ctx, ctx.TxSum())
 			if err == nil {
 				for _, l := range syntheticReceipt.Logs {
@@ -297,10 +302,8 @@ func (k *Keeper) applyEVMMessageWithTracing(
 	cfg := types.DefaultChainConfig().EthereumConfig(k.ChainID(ctx))
 	txCtx := core.NewEVMTxContext(msg)
 	evmHooks := evmtracers.GetCtxEthTracingHooks(ctx)
-	evmInstance := vm.NewEVM(*blockCtx, txCtx, stateDB, cfg, vm.Config{
-		Tracer: evmHooks,
-	})
-	stateDB.SetEVM(evmInstance)
+	evmInstance := vm.NewEVM(*blockCtx, txCtx, stateDB, cfg, vm.Config{Tracer: evmHooks}, k.customPrecompiles)
+
 	stateDB.SetLogger(evmHooks)
 
 	if onStart != nil {