+++ x/evm/module_test.go
@@ -140,10 +140,7 @@ func TestABCI(t *testing.T) {
 	require.Equal(t, receipt.BlockNumber, uint64(ctx.BlockHeight()))
 	require.Equal(t, receipt.VmError, "test error")
 
-	// fourth block with locked tokens in coinbase address
-	balanceChanges = make([]evmBalanceChange, 0)
-	ctx = addTestBalanceChangeTracerToCtx(ctx, &balanceChanges)
-
+	// disallow creating vesting account for coinbase address
 	m.BeginBlock(ctx, abci.RequestBeginBlock{})
 	coinbase := state.GetCoinbaseAddress(2)
 	vms := vesting.NewMsgServerImpl(*k.AccountKeeper(), k.BankKeeper())
@@ -153,25 +150,7 @@ func TestABCI(t *testing.T) {
 		Amount:      sdk.NewCoins(sdk.NewCoin("usei", sdk.OneInt())),
 		EndTime:     math.MaxInt64,
 	})
-	require.Nil(t, err)
-	s = state.NewDBImpl(ctx.WithTxIndex(2), k, false)
-	s.SubBalance(evmAddr1, big.NewInt(2000000000000), ethtracing.BalanceChangeUnspecified)
-	s.AddBalance(evmAddr2, big.NewInt(1000000000000), ethtracing.BalanceChangeUnspecified)
-	s.AddBalance(feeCollectorAddr, big.NewInt(1000000000000), ethtracing.BalanceChangeUnspecified)
-	surplus, err = s.Finalize()
-	require.Nil(t, err)
-	k.AppendToEvmTxDeferredInfo(ctx.WithTxIndex(2), ethtypes.Bloom{}, common.Hash{}, surplus)
-	k.SetTxResults([]*abci.ExecTxResult{{Code: 0}, {Code: 0}, {Code: 0}})
-	k.SetMsgs([]*types.MsgEVMTransaction{nil, nil, {}})
-	require.Equal(t, sdk.OneInt(), k.BankKeeper().SpendableCoins(ctx, coinbase).AmountOf("usei"))
-	m.EndBlock(ctx, abci.RequestEndBlock{}) // should not crash
-	require.Equal(t, sdk.OneInt(), k.BankKeeper().GetBalance(ctx, coinbase, "usei").Amount)
-	require.Equal(t, sdk.ZeroInt(), k.BankKeeper().SpendableCoins(ctx, coinbase).AmountOf("usei"))
-
-	require.Equal(t, 1, len(balanceChanges))
-	require.Equal(t, []evmBalanceChange{
-		{"2000000000000", "3000000000000", ethtracing.BalanceIncreaseRewardTransactionFee},
-	}, balanceChanges, "balance changes do not match, actual are:\n\n%s", balanceChangesValues(balanceChanges))
+	require.NotNil(t, err)
 }
 
 func TestAnteSurplus(t *testing.T) {